refactor(dto): extract skill IPC contracts
- move Skill request and response types into the DTO layer

- keep command handlers focused on transport delegation
