+++ src-tauri/src/lib.rs
@@ -27,6 +27,7 @@ mod db {
 mod dto {
     pub(crate) mod agent;
     pub(crate) mod comparison;
+    pub(crate) mod skill;
     pub(crate) mod workspace;
 }
 mod domain {
